requirements.txt
kivy>=1.11.1
plyer
kivymd_extensions.sweetalert